Build with Maven and Tycho
This commit adds a setup to build the Xturtle editor with Maven
and the Tycho plugin (call `mvn clean install` in the repo root).

The build creates a p2 Eclipse repository in the `updatesite`
project (under the `target` directory). This p2 repository makes
the editor consumable by other Eclipse features, e.g. by pointing
to it from a target platform definition in another Tycho build.